pyproject.toml
[project]
name = "expense-tracker-server"
version = "0.1.0"
description = "Lets AI assistants manage your personal expenses"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=3.2.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
